+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.0-beta.8
 
 * Support version 1.0.0-beta.11 of the Sass embedded protocol:
-  * Set CanonicalizeRequest.from_import.
+  * Support `VersionRequest` and `VersionResponse`.
+  * Set `CanonicalizeRequest.from_import`.
 
 * Properly throw errors for range checks for colors.